Fix subtest and round for python 2 compatibility

+++ torch/nn/utils/prune.py
@@ -1255,7 +1255,7 @@ def _compute_nparams_toprune(amount, tensor_size):
     if isinstance(amount, numbers.Integral):
         return amount
     else:
-        return round(amount * tensor_size)
+        return int(round(amount * tensor_size))  # int needed for Python 2
 
 
 def _validate_pruning_dim(t, dim):
